Overworld1
I put all the files containing the Overworld in a seperate project
folder since Github is not allowing me to merge the files. In the
future, we should all have our own branches so as to not create problems
in the Git master branch, which from past experience does not like
having multiple people pushing files into it. Merging seems to work
better.

Related to spawning:
If you look at the scene Overworld1, you will notice that all the tiles
are present in the hierarchy. In the future I want to find a solution
that is more memory efficient and that provides less clutter in the
hierarchy. For now, I think it might be a better idea to just put other
game objects in a parent class gameobject to organize the hierarchy
easier.

The benefit of having all the tiles available on the scene and in the
hierarchy is that they are easy to interact with. A solution for
spawning is to take the desired tiles for spawning and place them in
lists organized by what will be spawning at these tiles. For heros,
there could be a few public lists that represent the different locations
where heros can spawn based on user input. The corresponding tiles would
then be dragged into these lists on the unity editor and have attached
spawn scripts to spawn a specific gameobject (the hero in this case).
Monster spawning can also happen similarly.